Refactor BalancedAllocator.Balancer to LocalShardsBalancer
